Move `ValidationError` to `validation`
