+++ crates/fj-core/src/objects/any_object.rs
@@ -4,7 +4,8 @@ use crate::{
         Surface, Vertex,
     },
     storage::{Handle, HandleWrapper, ObjectId},
-    validate::{Validate, ValidationConfig, ValidationError},
+    validate::{Validate, ValidationConfig},
+    validation::ValidationError,
 };
 
 macro_rules! any_object {
