+++ crates/fj-core/src/validate/curve.rs
@@ -1,6 +1,6 @@
-use crate::objects::Curve;
+use crate::{objects::Curve, validation::ValidationError};
 
-use super::{Validate, ValidationConfig, ValidationError};
+use super::{Validate, ValidationConfig};
 
 impl Validate for Curve {
     fn validate_with_config(
