+++ crates/fj-core/src/validate/face.rs
@@ -1,8 +1,8 @@
 use fj_math::Winding;
 
-use crate::objects::Face;
+use crate::{objects::Face, validation::ValidationError};
 
-use super::{Validate, ValidationConfig, ValidationError};
+use super::{Validate, ValidationConfig};
 
 impl Validate for Face {
     fn validate_with_config(
@@ -94,7 +94,8 @@ mod tests {
             reverse::Reverse,
             update::{UpdateCycle, UpdateFace, UpdateRegion},
         },
-        validate::{FaceValidationError, Validate, ValidationError},
+        validate::{FaceValidationError, Validate},
+        validation::ValidationError,
         Core,
     };
